Added Additions & Changes for Season 12
Clarified captains only have to submit match results if the match isn't ticketed. Added entire section on participant conduct. Clarified there are four sources of free agents (added the free agent sign-up for seasons it is utilized). Clarified stand-in rules regarding number of times they can be used in the season and playoffs. Added rule to clearly state you cannot use a stand-in if any 5 of the actual team is online. Changed required stream delay to 5 minutes and clarified that admins & authorized casters are able to spectate in game.